requirements.txt
bitstruct==6.0.0
pytest==4.3.0
pyusb==1.0.2
